Merge Develop to Master

+++ src/main/java/de/beckerdd/bennet/minecraft/technicserver/TechnicAPI.java
@@ -1,20 +1,29 @@
 package de.beckerdd.bennet.minecraft.technicserver;
 
-import de.beckerdd.bennet.minecraft.technicserver.Helper.ClientMod;
-import de.beckerdd.bennet.minecraft.technicserver.Helper.Logging;
-import de.beckerdd.bennet.minecraft.technicserver.minecraftforge.installer.InstallerAction;
-import de.beckerdd.bennet.minecraft.technicserver.minecraftforge.installer.ServerInstall;
-import de.beckerdd.bennet.minecraft.technicserver.minecraftforge.installer.VersionInfo;
+import de.beckerdd.bennet.minecraft.technicserver.config.StaticConfig;
+import de.beckerdd.bennet.minecraft.technicserver.config.UserConfig;
+import de.beckerdd.bennet.minecraft.technicserver.technic.Modpack;
+import de.beckerdd.bennet.minecraft.technicserver.util.ANSIColor;
+import de.beckerdd.bennet.minecraft.technicserver.util.ClientMod;
+import de.beckerdd.bennet.minecraft.technicserver.util.ForgeInstaller;
+import de.beckerdd.bennet.minecraft.technicserver.util.Logging;
 import org.apache.commons.io.FileUtils;
+import org.piwik.java.tracking.CustomVariable;
+import org.piwik.java.tracking.PiwikRequest;
+import org.piwik.java.tracking.PiwikTracker;
 
 import java.io.*;
-import java.net.*;
-import java.nio.file.*;
+import java.net.JarURLConnection;
+import java.net.MalformedURLException;
+import java.net.URL;
+import java.net.URLConnection;
+import java.nio.file.Paths;
+import java.util.concurrent.atomic.AtomicInteger;
 import java.util.jar.JarEntry;
 import java.util.jar.JarFile;
 
 
-/**
+/*
  * Created by bennet on 8/7/17.
  *
  * technicserver - run modpacks from technicpack.net as server with ease.
@@ -60,10 +69,7 @@ public class TechnicAPI {
             throw new MalformedURLException("Not an API URL");
         }
 
-        if(!(state = new File("modpack.state")).exists()) {
-            Logging.logDebug("No State File Found");
-            modpack = new Modpack(conn.getInputStream());
-        }else{
+        if ((state = new File("modpack.state")).exists()) {
             Logging.logDebug("Loading old State");
             try {
                 FileInputStream fis = new FileInputStream(state);
@@ -74,11 +80,82 @@ public class TechnicAPI {
                 fis.close();
 
                 modpack.update(conn.getInputStream());
-            }catch (Exception ignore){
+            } catch (Exception ignore) {
                 Logging.logErr("STATE FILE INVALID! PLEASE RESTART!");
                 FileUtils.forceDelete(state);
                 System.exit(1);
             }
+        } else {
+            Logging.logDebug("No State File Found");
+            modpack = new Modpack(conn.getInputStream());
+        }
+    }
+
+    /**
+     * Report Statisics to Piwik instance
+     * @param runtime Loadup time
+     */
+    public void runAnalytics(String runtime) {
+        runAnalytics(runtime,false);
+    }
+
+    /**
+     * Report Statisics to Piwik instance
+     * @param runtime Loadup time
+     * @param force override user config
+     */
+    public void runAnalytics(String runtime, boolean force) {
+        if (!UserConfig.isDisableAnalytics() || force) {
+            PiwikRequest piwikRequest;
+            AtomicInteger ruleCounter = new AtomicInteger(0);
+
+            try {
+                piwikRequest = new PiwikRequest(2, new URL(StaticConfig.PIWIK_PROPERTY));
+            } catch (MalformedURLException e) {
+                return;
+            }
+            PiwikTracker piwikTracker = new PiwikTracker(StaticConfig.PIWIK_URL);
+
+            if (Main.class.getPackage().getImplementationVersion() == null) {
+                if (System.getenv("TRAVIS") != null && System.getenv("TRAVIS")
+                        .equalsIgnoreCase("true")) {
+                    piwikRequest.addCustomTrackingParameter("implementation-version",
+                            "TRAVIS #" + System.getenv("TRAVIS_BUILD_NUMBER") +
+                                    " @ " + System.getenv("TRAVIS_COMMIT"));
+
+                    piwikRequest.setPageCustomVariable(new CustomVariable("implementation-version", "TRAVIS"),
+                            ruleCounter.incrementAndGet());
+                } else {
+                    piwikRequest.addCustomTrackingParameter("implementation-version", "DEBUG-RUN");
+                    piwikRequest.setPageCustomVariable(new CustomVariable("implementation-version", "DEBUG-RUN"),
+                            ruleCounter.incrementAndGet());
+                }
+            } else {
+                piwikRequest.addCustomTrackingParameter("implementation-version",
+                        Main.class.getPackage().getImplementationVersion());
+                piwikRequest.setPageCustomVariable(new CustomVariable("implementation-version",
+                        Main.class.getPackage().getImplementationVersion()), ruleCounter.incrementAndGet());
+            }
+
+            piwikRequest.addCustomTrackingParameter("modpack-url", UserConfig.getUrl());
+            piwikRequest.setPageCustomVariable(new CustomVariable("modpack-url", UserConfig.getUrl()),
+                    ruleCounter.incrementAndGet());
+            piwikRequest.addCustomTrackingParameter("modpack-desired-version", UserConfig.getBuild());
+            piwikRequest.setPageCustomVariable(new CustomVariable("modpack-desired-version", UserConfig.getBuild()),
+                    ruleCounter.incrementAndGet());
+            piwikRequest.addCustomTrackingParameter("modpack-autoupdate", UserConfig.isAutoupdate());
+            piwikRequest.setPageCustomVariable(new CustomVariable("modpack-autoupdate",
+                    UserConfig.isAutoupdate() ? "true" : "false"), ruleCounter.incrementAndGet());
+            piwikRequest.addCustomTrackingParameter("modpack-name", modpack.getDisplayName());
+            piwikRequest.setPageCustomVariable(new CustomVariable("modpack-name", modpack.getDisplayName()),
+                    ruleCounter.incrementAndGet());
+            piwikRequest.addCustomTrackingParameter("modpack-setuptime", runtime);
+            piwikRequest.setPageCustomVariable(new CustomVariable("modpack-setuptime", runtime),
+                    ruleCounter.incrementAndGet());
+
+            try {
+                piwikTracker.sendRequest(piwikRequest);
+            } catch (IOException ignore) {  }
         }
     }
 
@@ -88,17 +165,17 @@ public class TechnicAPI {
      */
     public void downloadAndInstallModpack() throws IOException {
         modpack.downloadAll().extractAll();
-        try{
+        try {
             Logging.log("Downloading Server Icon");
             modpack.getIcon().download("server-icon.png");
-        }catch (MalformedURLException ignore){
+        } catch (MalformedURLException ignore) {
             Logging.logErr("Icon Download Failed");
         }
     }
 
     /**
      * Convert the Client Modpack to a Server Compatible one, by installing the Forge ModLoader
-     * @throws IOException
+     * @throws IOException Failed Downloading Forge or moving modpack.jar
      */
     public void convertServer() throws IOException {
         URL url = new URL(
@@ -107,34 +184,33 @@ public class TechnicAPI {
         JarFile jarFile = conn.getJarFile();
         JarEntry jarEntry = conn.getJarEntry();
 
-        VersionInfo.overriddenMinecraftVersion = modpack.getMinecraft().toString();
-        VersionInfo.LazyInit(jarFile.getInputStream(jarEntry));
-        ServerInstall.headless = true;
-        if (!InstallerAction.SERVER.run(new File("."), s -> true)) {
-            Logging.logErr("There was an error during server installation");
-        } else {
-            Files.copy(
-                    FileSystems.getDefault().getPath("bin","modpack.jar"),
-                    FileSystems.getDefault().getPath("modpack.jar"), StandardCopyOption.REPLACE_EXISTING);
-            Logging.log("Forge Mod Loader successfully installed");
-        }
+        ForgeInstaller.installForge(jarFile.getInputStream(jarEntry));
+        FileUtils.copyFile(new File("bin/modpack.jar"), new File("modpack.jar"));
+        modpack.getMinecraft().download();
+
+        Logging.log("Forge Mod Loader successfully installed");
     }
 
     /**
      * Delete Client-Only Mods which may cause trouble with the Server
      */
     public void cleanClientMods() {
-        for(File child : new File("mods/").listFiles()){
-            if(ClientMod.isClientMod(child.getName())){
-                Logging.log("Deleting Clientmod-File " + child.getName());
-                if(child.isDirectory()){
-                    try {
-                        FileUtils.deleteDirectory(child);
-                    } catch (IOException e) {
-                        e.printStackTrace();
+        File[] modsDir = new File("mods/").listFiles();
+        if (modsDir == null) {
+            Logging.logErr("Ohh, your seams mods folder empty o.Ô");
+        } else {
+            for (File child : modsDir) {
+                if (ClientMod.isClientMod(child.getName())) {
+                    Logging.log("Deleting Clientmod-File " + child.getName());
+                    if (child.isDirectory()) {
+                        try {
+                            FileUtils.deleteDirectory(child);
+                        } catch (IOException e) {
+                            e.printStackTrace();
+                        }
+                    } else {
+                        child.delete();
                     }
-                }else {
-                    boolean b = child.delete();
                 }
             }
         }
@@ -162,5 +238,28 @@ public class TechnicAPI {
         oos.flush();
         oos.close();
         fos.close();
+    }
+
+    /**
+     * Check for Modpack update
+     * @param force override user config
+     * @throws IOException Download, Extract or Convert failed
+     * @throws InterruptedException sleep failed
+     */
+    public void updatePack(boolean force) throws IOException, InterruptedException {
+        if (UserConfig.isAutoupdate() || force ||
+                this.getModpack().getState() == Modpack.State.NOT_INSTALLED ) {
+            this.downloadAndInstallModpack();
+            this.convertServer();
+            this.cleanClientMods();
+        } else {
+            Logging.log(ANSIColor.ANSI_RED + ANSIColor.ANSI_YELLOW_BACKGROUND +
+                    "MODPACK UPDATE AVAILABLE, BUT AUTOUPDATE IS DISABLED. " +
+                    "START WITH \"update\" AS PARAMETER OR SET AUTOUPDATE TO \"yes\"" +
+                    ANSIColor.ANSI_RESET);
+
+            Logging.logDebug("sleeping 5 sec.");
+            Thread.sleep(5000);
+        }
     }
 }
